chore: Organize module directory structure
