+++ family/src/main/java/com/oing/repository/FamilyRepository.java
@@ -1,8 +1,7 @@
 package com.oing.repository;
 
-import com.oing.domain.Family;
+import com.oing.domain.model.Family;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface FamilyRepository extends JpaRepository<Family, String> {
 }
