+++ member/src/main/java/com/oing/exception/TokenNotValidException.java
@@ -1,4 +1,7 @@
-package com.oing.domain.exception;
+package com.oing.exception;
+
+import com.oing.domain.exception.DomainException;
+import com.oing.domain.exception.ErrorCode;
 
 /**
  * no5ing-server
