+++ member/src/main/java/com/oing/service/TokenGenerator.java
@@ -3,7 +3,7 @@ package com.oing.service;
 import com.oing.domain.SocialLoginProvider;
 import com.oing.domain.Token;
 import com.oing.domain.TokenPair;
-import com.oing.domain.TokenType;
+import com.oing.exception.TokenNotValidException;
 
 /**
  * no5ing-server
@@ -30,7 +30,7 @@ public interface TokenGenerator {
     /**
      * 토큰의 데이터를 추출합니다.
      * 유효하지 않은 경우, TokenNotValidException 이 발생합니다.
-     * @throws com.oing.domain.exception.TokenNotValidException 토큰이 유효하지 않은 경우
+     * @throws TokenNotValidException 토큰이 유효하지 않은 경우
      * @param token JWT 토큰
      * @return 토큰 데이터 DTO
      */
