+++ member/src/test/java/com/oing/domain/exception/TokenNotValidExceptionTest.java
@@ -1,5 +1,6 @@
 package com.oing.domain.exception;
 
+import com.oing.exception.TokenNotValidException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
